# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/transactions/borrow": {
      "post": {
        "operationId": "TransactionsController_borrow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Transaction"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Transaction"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pinjam buku",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/transactions/return/{barcode}": {
      "patch": {
        "operationId": "TransactionsController_returnBook",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Transaction"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Transaction"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kembalikan buku berdasarkan barcode",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci pencarian",
            "schema": {
              "example": "search term",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan jenis aksi audit log",
            "schema": {
              "example": "CREATE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Transaction"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Transaction"
                          }
                        }
                      ],
                      "nullable": true
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "total_items": {
                          "type": "number",
                          "example": 0
                        },
                        "total_pages": {
                          "type": "number",
                          "example": 0
                        },
                        "has_next_page": {
                          "type": "boolean",
                          "example": false
                        },
                        "has_prev_page": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Riwayat semua transaksi (paginated)",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/transactions/my-history": {
      "get": {
        "operationId": "TransactionsController_myHistory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci pencarian",
            "schema": {
              "example": "search term",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan jenis aksi audit log",
            "schema": {
              "example": "CREATE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Transaction"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Transaction"
                          }
                        }
                      ],
                      "nullable": true
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "total_items": {
                          "type": "number",
                          "example": 0
                        },
                        "total_pages": {
                          "type": "number",
                          "example": 0
                        },
                        "has_next_page": {
                          "type": "boolean",
                          "example": false
                        },
                        "has_prev_page": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Riwayat transaksi user yang login",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Digital API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Alternatif untuk cookie: tempel access_token hasil login."
      },
      "access_token": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly cookie access token, diset otomatis oleh POST /auth/login."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "identification_number": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "role": {
            "enum": [
              "SUPER_ADMIN",
              "STAFF",
              "USER"
            ],
            "type": "string"
          },
          "category": {
            "enum": [
              "STUDENT",
              "LECTURER",
              "LIBRARY_STAFF",
              "PUBLIC"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "identification_number",
          "email",
          "full_name",
          "role",
          "category",
          "created_at",
          "updated_at"
        ]
      },
...
      "BookItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "barcode": {
            "type": "string"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "BORROWED",
              "LOST",
              "DAMAGED"
            ],
            "type": "string"
          },
          "condition": {
            "enum": [
              "GOOD",
              "SLIGHTLY_DAMAGED",
              "HEAVILY_DAMAGED"
            ],
            "type": "string"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "added_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "barcode",
          "status",
          "condition",
          "book",
          "added_at",
          "updated_at"
        ]
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "bookItem": {
            "$ref": "#/components/schemas/BookItem"
          },
          "borrowed_at": {
            "format": "date-time",
            "type": "string"
          },
          "due_date": {
            "format": "date-time",
            "type": "string"
          },
          "returned_at": {
            "format": "date-time",
            "type": "string"
          },
          "fine_amount": {
            "type": "number"
          },
          "status": {
            "enum": [
              "BORROWED",
              "RETURNED",
              "OVERDUE"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "bookItem",
          "borrowed_at",
          "due_date",
          "returned_at",
          "fine_amount",
          "status"
        ]
      },
      "BorrowBookDto": {
        "type": "object",
        "properties": {
          "barcode": {
            "type": "string",
            "example": "B001",
            "description": "Barcode fisik buku"
          }
        },
        "required": [
          "barcode"
        ]
      },
...
    }
  }
}